refactor: Streamline workout instances context and related components
- Refactored WorkoutInstancesContext to utilize a centralized hooks structure, improving organization and maintainability.
- Updated imports in various components to reflect the new hooks directory, enhancing clarity and reducing redundancy.
- Removed obsolete context definitions and custom hooks related to workout instances, streamlining the codebase for better performance and readability.
