+++ src/hooks/index.ts
@@ -23,3 +23,6 @@ export {
   useUserAccessContext,
   MeteredFeature,
 } from "./useUserAccess";
+export * from "./useWorkoutInstances";
+export { usePusherBeamsInterests } from "./usePusherBeamsInterests";
+export { usePusherEvent } from "./usePusherEvent";
