+++ src/modules/dashboard/workouts/components/WorkoutComponents.tsx
@@ -1,21 +1,7 @@
 "use client";
 
 import { Exercise } from "@/domain/entities/generatedWorkout";
-
-// Helper function to format time in seconds to minutes:seconds
-export const formatTime = (seconds?: number): string => {
-  if (
-    seconds === undefined ||
-    seconds === null ||
-    isNaN(seconds) ||
-    seconds < 0
-  ) {
-    return "0:00";
-  }
-  const mins = Math.floor(seconds / 60);
-  const secs = Math.floor(seconds % 60);
-  return `${mins}:${secs.toString().padStart(2, "0")}`;
-};
+import { formatTime } from "../utils/workouts.func";
 
 // Exercise component to display a single exercise
 export const ExerciseCard = ({ exercise }: { exercise: Exercise }) => {
