+++ src/modules/dashboard/workouts/components/WorkoutInstanceModal.tsx
@@ -1,16 +1,16 @@
-import { useEffect, useCallback, useState } from "react";
 import { WorkoutInstance } from "@/domain/entities/workoutInstance";
 import {
-  X,
-  Clock,
+  Activity,
   Calendar,
   CheckCircle,
   Circle,
-  Activity,
-  Target,
-  StickyNote,
+  Clock,
   ExternalLink,
+  StickyNote,
+  Target,
+  X,
 } from "lucide-react";
+import { useCallback, useEffect } from "react";
 import { formatDate } from "../utils/workoutHistoryUtils";
 
 export interface WorkoutInstanceModalProps {
@@ -287,31 +287,4 @@ export default function WorkoutInstanceModal({
       </div>
     </div>
   );
-}
-
-/**
- * Hook to manage workout instance modal state
- */
-export function useWorkoutInstanceModal(initialOpen = false) {
-  const [isOpen, setIsOpen] = useState(initialOpen);
-  const [workoutInstance, setWorkoutInstance] =
-    useState<WorkoutInstance | null>(null);
-
-  const openModal = useCallback((instance: WorkoutInstance) => {
-    setWorkoutInstance(instance);
-    setIsOpen(true);
-  }, []);
-
-  const closeModal = useCallback(() => {
-    setIsOpen(false);
-    // Keep the instance data for a smooth close animation
-    setTimeout(() => setWorkoutInstance(null), 200);
-  }, []);
-
-  return {
-    isOpen,
-    workoutInstance,
-    openModal,
-    closeModal,
-  };
 }
